fix: use absolute import for config in app factory
