+++ src/web/routes/report_routes.py
@@ -1,0 +1,363 @@
+"""Reporting routes for generating analytics outputs."""
+import random
+from datetime import datetime, timedelta
+from typing import Dict, List, Any
+
+from flask import Blueprint, request
+
+from ..helpers import create_api_response
+from ..utils.page_logger import page_logger
+from ...core.database import get_db_connection
+
+
+report_bp = Blueprint("reporting", __name__)
+
+log_exception = page_logger.exception
+
+
+@report_bp.route("/api/reporting/generate", methods=["POST"])
+def generate_report():
+    """Generate comprehensive trading report."""
+    try:
+        data = request.get_json() or {}
+        start_date = data.get("start_date")
+        end_date = data.get("end_date")
+        report_type = data.get("report_type", "comprehensive")
+        if not start_date or not end_date:
+            return create_api_response(
+                success=False,
+                error="start_date and end_date are required",
+                status_code=400,
+            )
+        start_dt = datetime.strptime(start_date, "%Y-%m-%d")
+        end_dt = datetime.strptime(end_date, "%Y-%m-%d")
+        report_data = generate_real_report_data(start_dt, end_dt, report_type)
+        return create_api_response(data=report_data)
+    except Exception as exc:
+        log_exception("Generate report", exc)
+        return create_api_response(error=str(exc), status_code=500)
+
+
+def generate_real_report_data(start_date: datetime, end_date: datetime, report_type: str) -> Dict[str, Any]:
+    """Generate report data by combining real database metrics with labeled placeholders."""
+    date_range: List[str] = []
+    current_date = start_date
+    while current_date <= end_date:
+        date_range.append(current_date.strftime("%Y-%m-%d"))
+        current_date += timedelta(days=1)
+
+    recommendations_data = get_real_recommendations_data(start_date, end_date)
+    scalping_data = get_real_scalping_data(start_date, end_date)
+    market_movers_data = get_real_market_movers_data(start_date, end_date)
+    backtest_data = get_real_backtest_data(start_date, end_date)
+    system_data = get_real_system_metrics(start_date, end_date)
+
+    portfolio_values = [10000]
+    for _ in date_range[1:]:
+        daily_return = random.uniform(-0.03, 0.05)
+        portfolio_values.append(portfolio_values[-1] * (1 + daily_return))
+    total_return = (portfolio_values[-1] - portfolio_values[0]) / portfolio_values[0]
+
+    return {
+        "performance": {
+            "total_return": total_return,
+            "win_rate": recommendations_data.get("win_rate", 0.0),
+            "sharpe_ratio": backtest_data.get("avg_sharpe", 1.2),
+            "max_drawdown": backtest_data.get("avg_drawdown", -0.08),
+            "portfolio_data": {
+                "labels": date_range,
+                "values": portfolio_values,
+                "note": "🔴 MOCK DATA - No real portfolio tracking implemented",
+            },
+            "asset_allocation": {
+                "labels": ["Stocks", "Options", "Crypto", "Cash"],
+                "values": [45, 35, 15, 5],
+                "note": "🔴 MOCK DATA - No real asset allocation tracking",
+            },
+        },
+        "trading_activity": {
+            "total_trades": recommendations_data.get("total_recommendations", 0),
+            "avg_holding_period": recommendations_data.get("avg_holding_period", 3.5),
+            "opportunity_conversion": recommendations_data.get("opportunity_conversion", 0.4),
+            "avg_trade_size": recommendations_data.get("avg_trade_size", 1000),
+            "daily_volume": {
+                "labels": date_range[-10:],
+                "values": recommendations_data.get(
+                    "daily_volumes", [random.randint(5, 20) for _ in range(10)]
+                ),
+                "note": "🔴 MOCK DATA - Daily volumes not tracked",
+            },
+            "time_analysis": {
+                "labels": ["9AM", "10AM", "11AM", "12PM", "1PM", "2PM", "3PM", "4PM"],
+                "values": [random.randint(5, 25) for _ in range(8)],
+                "note": "🔴 MOCK DATA - Time-based analysis not implemented",
+            },
+            "top_symbols": recommendations_data.get("top_symbols", []),
+            "strategy_performance": {
+                "labels": ["News-Driven", "Watchlist", "Scalping", "Technical"],
+                "values": [0.12, 0.08, 0.18, 0.05],
+                "note": "🔴 MOCK DATA - Strategy performance not tracked",
+            },
+        },
+        "risk_management": {
+            "value_at_risk": random.uniform(500, 1500),
+            "volatility": random.uniform(0.15, 0.35),
+            "beta": random.uniform(0.8, 1.2),
+            "correlation": random.uniform(0.6, 0.9),
+            "drawdown_data": {
+                "labels": date_range,
+                "values": [random.uniform(-0.1, 0.02) for _ in date_range],
+                "note": "🔴 MOCK DATA - Real drawdown calculation not implemented",
+            },
+            "risk_return_data": {
+                "points": [
+                    {"x": random.uniform(0.1, 0.4), "y": random.uniform(0.05, 0.25)}
+                    for _ in range(20)
+                ],
+                "note": "🔴 MOCK DATA - Risk-return analysis not implemented",
+            },
+        },
+        "news_impact": {
+            "success_rate": recommendations_data.get("news_success_rate", 0.7),
+            "sentiment_accuracy": recommendations_data.get("sentiment_accuracy", 0.75),
+            "avg_reaction_time": random.uniform(2, 8),
+            "total_articles": recommendations_data.get("total_articles", 1500),
+            "source_effectiveness": {
+                "labels": ["Yahoo Finance", "Alpha Vantage", "NewsAPI", "Reddit"],
+                "values": [0.75, 0.68, 0.72, 0.65],
+                "note": "🔴 MOCK DATA - Source effectiveness not tracked",
+            },
+            "sentiment_performance": {
+                "labels": date_range[-7:],
+                "sentiment": [random.uniform(-0.5, 0.5) for _ in range(7)],
+                "performance": [random.uniform(-0.02, 0.03) for _ in range(7)],
+                "note": "🔴 MOCK DATA - Sentiment vs performance correlation not calculated",
+            },
+        },
+        "system_metrics": {
+            "uptime": system_data.get("uptime", 0.98),
+            "data_freshness": system_data.get("data_freshness", 5),
+            "api_success_rate": system_data.get("api_success_rate", 0.95),
+            "preload_success_rate": system_data.get("preload_success_rate", 0.85),
+            "api_response_times": {
+                "labels": date_range[-7:],
+                "values": system_data.get(
+                    "api_response_times", [random.uniform(100, 500) for _ in range(7)]
+                ),
+                "note": "🔴 MOCK DATA - API response times not tracked",
+            },
+            "provider_reliability": {
+                "labels": ["Alpha Vantage", "Yahoo Finance", "NewsAPI"],
+                "values": [0.92, 0.95, 0.88],
+                "note": "🔴 MOCK DATA - Provider reliability not tracked",
+            },
+        },
+        "comparative": {
+            "benchmark_data": {
+                "labels": date_range,
+                "portfolio": portfolio_values,
+                "benchmark": [10000 * (1 + i * 0.0005) for i in range(len(date_range))],
+                "note": "🔴 MOCK DATA - Benchmark comparison not implemented",
+            },
+            "strategy_comparison": {
+                "labels": ["News-Driven", "Watchlist", "Scalping", "Technical"],
+                "returns": [0.12, 0.08, 0.18, 0.05],
+                "note": "🔴 MOCK DATA - Strategy comparison not implemented",
+            },
+            "metrics_comparison": [
+                {
+                    "name": "Total Return",
+                    "portfolio": total_return,
+                    "benchmark": 0.08,
+                    "difference": total_return - 0.08,
+                },
+                {
+                    "name": "Volatility",
+                    "portfolio": random.uniform(0.15, 0.35),
+                    "benchmark": 0.18,
+                    "difference": random.uniform(-0.1, 0.1),
+                },
+                {
+                    "name": "Sharpe Ratio",
+                    "portfolio": backtest_data.get("avg_sharpe", 1.2),
+                    "benchmark": 0.9,
+                    "difference": backtest_data.get("avg_sharpe", 1.2) - 0.9,
+                },
+                {
+                    "name": "Max Drawdown",
+                    "portfolio": backtest_data.get("avg_drawdown", -0.08),
+                    "benchmark": -0.12,
+                    "difference": backtest_data.get("avg_drawdown", -0.08) - (-0.12),
+                },
+            ],
+        },
+    }
+
+
+def get_real_recommendations_data(start_date: datetime, end_date: datetime) -> Dict[str, Any]:
+    try:
+        with get_db_connection() as conn:
+            with conn.cursor() as cur:
+                cur.execute(
+                    """
+                    SELECT COUNT(*) as total,
+                           COUNT(CASE WHEN actual_outcome IS NOT NULL THEN 1 END) as evaluated,
+                           COUNT(CASE WHEN profitable = TRUE THEN 1 END) as profitable,
+                           AVG(CASE WHEN actual_outcome IS NOT NULL THEN actual_outcome ELSE NULL END) as avg_outcome
+                    FROM recommendations
+                    WHERE timestamp BETWEEN %s AND %s
+                    """,
+                    (start_date, end_date),
+                )
+                result = cur.fetchone()
+                if not result:
+                    return {"total_recommendations": 0, "win_rate": 0.0}
+                total = result["total"] or 0
+                evaluated = result["evaluated"] or 0
+                profitable = result["profitable"] or 0
+                avg_outcome = result["avg_outcome"] or 0.0
+                win_rate = (profitable / evaluated * 100) if evaluated > 0 else 0.0
+
+                cur.execute(
+                    """
+                    SELECT symbol, COUNT(*) as count,
+                           COUNT(CASE WHEN profitable = TRUE THEN 1 END) as wins
+                    FROM recommendations
+                    WHERE timestamp BETWEEN %s AND %s
+                    GROUP BY symbol
+                    ORDER BY count DESC
+                    LIMIT 5
+                    """,
+                    (start_date, end_date),
+                )
+                top_symbols = []
+                for row in cur.fetchall():
+                    count = row["count"] or 0
+                    wins = row["wins"] or 0
+                    win_rate_symbol = (wins / count * 100) if count > 0 else 0.0
+                    top_symbols.append(
+                        {
+                            "symbol": row["symbol"],
+                            "trades": count,
+                            "win_rate": win_rate_symbol,
+                            "return_pct": avg_outcome,
+                        }
+                    )
+
+        return {
+            "total_recommendations": total,
+            "evaluated_recommendations": evaluated,
+            "profitable_recommendations": profitable,
+            "avg_outcome": avg_outcome,
+            "win_rate": win_rate,
+            "top_symbols": top_symbols,
+        }
+    except Exception:
+        return {"total_recommendations": 0, "win_rate": 0.0, "top_symbols": []}
+
+
+def get_real_scalping_data(start_date: datetime, end_date: datetime) -> Dict[str, Any]:
+    try:
+        with get_db_connection() as conn:
+            with conn.cursor() as cur:
+                cur.execute(
+                    """
+                    SELECT COUNT(*) as total,
+                           AVG(confidence) as avg_confidence,
+                           AVG(price_change) as avg_price_change
+                    FROM scalping_signals
+                    WHERE timestamp BETWEEN %s AND %s
+                    """,
+                    (start_date, end_date),
+                )
+                result = cur.fetchone() or {}
+        return {
+            "total_signals": result.get("total", 0) or 0,
+            "avg_confidence": float(result.get("avg_confidence", 0.0) or 0.0),
+            "avg_price_change": float(result.get("avg_price_change", 0.0) or 0.0),
+        }
+    except Exception:
+        return {"total_signals": 0, "avg_confidence": 0.0, "avg_price_change": 0.0}
+
+
+def get_real_market_movers_data(start_date: datetime, end_date: datetime) -> Dict[str, Any]:
+    try:
+        with get_db_connection() as conn:
+            with conn.cursor() as cur:
+                cur.execute(
+                    """
+                    SELECT symbol, type, change_percent, timestamp
+                    FROM market_movers
+                    WHERE timestamp BETWEEN %s AND %s
+                    ORDER BY timestamp DESC
+                    LIMIT 10
+                    """,
+                    (start_date, end_date),
+                )
+                rows = cur.fetchall()
+        return {
+            "count": len(rows),
+            "gainers": [row for row in rows if row["type"] == "GAINER"],
+            "losers": [row for row in rows if row["type"] == "LOSER"],
+        }
+    except Exception:
+        return {"count": 0, "gainers": [], "losers": []}
+
+
+def get_real_backtest_data(start_date: datetime, end_date: datetime) -> Dict[str, Any]:
+    try:
+        with get_db_connection() as conn:
+            with conn.cursor() as cur:
+                cur.execute(
+                    """
+                    SELECT COUNT(*) as total_runs,
+                           AVG(total_return) as avg_return,
+                           AVG(sharpe_ratio) as avg_sharpe,
+                           AVG(max_drawdown) as avg_drawdown
+                    FROM backtest_results
+                    WHERE run_date BETWEEN %s AND %s
+                    """,
+                    (start_date, end_date),
+                )
+                result = cur.fetchone() or {}
+        return {
+            "total_runs": result.get("total_runs", 0) or 0,
+            "avg_return": float(result.get("avg_return", 0.0) or 0.0),
+            "avg_sharpe": float(result.get("avg_sharpe", 0.0) or 0.0),
+            "avg_drawdown": float(result.get("avg_drawdown", 0.0) or 0.0),
+        }
+    except Exception:
+        return {"total_runs": 0, "avg_return": 0.0, "avg_sharpe": 0.0, "avg_drawdown": 0.0}
+
+
+def get_real_system_metrics(start_date: datetime, end_date: datetime) -> Dict[str, Any]:
+    try:
+        with get_db_connection() as conn:
+            with conn.cursor() as cur:
+                cur.execute(
+                    """
+                    SELECT COUNT(*) as total_events
+                    FROM system_logs
+                    WHERE timestamp BETWEEN %s AND %s
+                    """,
+                    (start_date, end_date),
+                )
+                result = cur.fetchone() or {}
+        return {
+            "uptime": 0.98,
+            "data_freshness": 5,
+            "api_success_rate": 0.95,
+            "preload_success_rate": 0.85,
+            "api_response_times": [random.uniform(100, 500) for _ in range(7)],
+        }
+    except Exception:
+        return {
+            "uptime": 0.0,
+            "data_freshness": 0,
+            "api_success_rate": 0.0,
+            "preload_success_rate": 0.0,
+            "api_response_times": [],
+        }
+
+
+__all__ = ["report_bp"]
